+++ utils/navController.ts
@@ -2,18 +2,9 @@
 import type { NavComponentWithProps, ComponentProps, NavComponent, NavOptions, TransitionDoneFn, ViewController } from "@ionic/core";
 import type { SvelteComponent } from "svelte";
 
-import { writable, get } from "svelte/store"
-
-// not sure if we need to stick to a writable - strictly spoken not necessary
-const ion_nav = writable<HTMLIonNavElement>(undefined)
-
-export function setActiveNavElement(element: HTMLIonNavElement) {
-    ion_nav.set(element);
-}
-
 const createHTMLCompFromSvelte = (
     component: new (...args: any) => SvelteComponent,
-    componentProps = {}
+    componentProps: { [key: string]: any } = {}
 ) => {
     const divWrapper = document.createElement('div');
     const contentID = 'id' + Date.now();
@@ -40,27 +31,27 @@ const createHTMLCompFromSvelte = (
 export const navController = {
 
     canGoBack: function (view?: ViewController) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         return _nav !== undefined ? _nav.canGoBack(view) : undefined;
     },
 
     getActive: function () {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         return _nav !== undefined ? _nav.getActive() : undefined;
     },
 
     getByIndex: function (index: number) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         return _nav !== undefined ? _nav.getByIndex(index) : undefined;
     },
 
     getPrevious: function (view?: ViewController) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         return _nav !== undefined ? _nav.getPrevious(view) : undefined;
     },
 
     insert: function <T extends NavComponent>(insertIndex: number, component: T, componentProps?: ComponentProps<T> | null, opts?: NavOptions | null, done?: TransitionDoneFn) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         // eslint-disable @typescript-eslint/ban-ts-comment
         //@ts-ignore
         const htmlcomponent = createHTMLCompFromSvelte(component, componentProps);
@@ -68,7 +59,7 @@ export const navController = {
     },
 
     insertPages: function (insertIndex: number, insertComponents: NavComponent[] | NavComponentWithProps[], opts?: NavOptions | null, done?: TransitionDoneFn) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         // eslint-disable @typescript-eslint/ban-ts-comment
         //@ts-ignore
         const htmlcomponents: NavComponent[] | NavComponentWithProps[] = insertComponents.map((component: NavComponent | NavComponentWithProps) => {
@@ -84,22 +75,22 @@ export const navController = {
     },
 
     pop: function (opts?: NavOptions | null, done?: TransitionDoneFn) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         return _nav !== undefined ? _nav.pop(opts, done) : undefined;
     },
 
     popTo: function (indexOrViewCtrl: number | ViewController, opts?: NavOptions | null, done?: TransitionDoneFn) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         return _nav !== undefined ? _nav.popTo(indexOrViewCtrl, opts, done) : undefined;
     },
 
     popToRoot: function (opts?: NavOptions | null, done?: TransitionDoneFn) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         return _nav !== undefined ? _nav.popToRoot(opts, done) : undefined;
     },
 
     push: function <T extends NavComponent>(component: T, componentProps?: ComponentProps<T>, opts?: NavOptions, done?: TransitionDoneFn) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         //@ts-ignore
         const htmlcomponent = createHTMLCompFromSvelte(component, componentProps);
@@ -107,12 +98,12 @@ export const navController = {
     },
 
     removeIndex: function (startIndex: number, removeCount?: number, opts?: NavOptions | null, done?: TransitionDoneFn) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         return _nav !== undefined ? _nav.removeIndex(startIndex, removeCount, opts, done) : undefined;
     },
 
     setPages: function (views: NavComponent[] | NavComponentWithProps[], opts?: NavOptions | null, done?: TransitionDoneFn) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         // eslint-disable @typescript-eslint/ban-ts-comment
         //@ts-ignore
         const htmlcomponents: NavComponent[] | NavComponentWithProps[] = views.map((component: NavComponent | NavComponentWithProps) => {
@@ -128,7 +119,7 @@ export const navController = {
     },
 
     setRoot: function <T extends NavComponent>(component: T, componentProps?: ComponentProps<T>, opts?: NavOptions, done?: TransitionDoneFn) {
-        const _nav = get(ion_nav);
+        const _nav = document.querySelector('ion-nav') as HTMLIonNavElement;
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         //@ts-ignore
         const htmlcomponent = createHTMLCompFromSvelte(component, componentProps);